- added a string formatter for the counter display - managed to get the vegetable object creation to work with money being set to a decimal value to display correctly - now working on creating an object array and passing the object from the button click to the array

+++ ContactlessSelfCheckout/ContactlessSelfCheckout/dbo.Table_Product.sql
@@ -2,7 +2,7 @@
     [Product_ID]       INT          NOT NULL,
     [Product_Name]     VARCHAR (50) NOT NULL,
     [Product_Category] VARCHAR (50) NOT NULL,
-    [Product_Price]    INT          DEFAULT ((0)) NOT NULL,
+    [Product_Price]    DECIMAL (10, 2) DEFAULT ((0)) NOT NULL,
     [Product_Stock]    INT          DEFAULT ((0)) NOT NULL,
     PRIMARY KEY CLUSTERED ([Product_ID] ASC)
 );
